System documentation
This is tge documentation covering the design, coding, testing and refining of teh system given in the project management scenario. Details of the system can be found in the READ.me file or you can go straight to the documentation and find it out yourself and view the actual system code and layout in the zip file.