System folder and file layout
The actual program for the code of the system can be found within the zip folder and is called assignment, the rest of the zip folder is the layout for modules and other libraries used in the system to operate everything used. You can find the spec of the system in the System Documentation.